@@ -20,12 +20,12 @@
         }
     },
     "require": {
-        "php": "^7.4",
+        "php": "^8.0",
         "oneup/flysystem-bundle": "^4.0",
-        "symfony/framework-bundle": "^5.0"
+        "symfony/framework-bundle": "^5.0 || ^6.0"
     },
     "require-dev": {
         "friendsofphp/php-cs-fixer": "^3.0",
-        "rector/rector": "^0.11.60"
+        "rector/rector": "^0.12"
     }
 }
