@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Subugoe\OaiBundle\Controller;
 
 use JMS\Serializer\SerializerInterface;
-use Subugoe\OaiBundle\Model\Identify\Identify;
 use Subugoe\OaiBundle\Service\OaiServiceInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -14,24 +13,16 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class OaiController extends AbstractController
 {
-    private \Subugoe\OaiBundle\Service\OaiServiceInterface $oaiService;
-    private \JMS\Serializer\SerializerInterface $serializer;
-
-    public function __construct(SerializerInterface $serializer, OaiServiceInterface $oaiService)
+    public function __construct(private SerializerInterface $serializer, private OaiServiceInterface $oaiService)
     {
-        $this->serializer = $serializer;
-        $this->oaiService = $oaiService;
     }
 
-    /**
-     * @Route("/oai2/verb/Identify")
-     */
+    #[Route(path: '/oai2/verb/Identify')]
     public function identify(Request $request)
     {
         $url = $request->getSchemeAndHttpHost().$request->getPathInfo();
         $oaiConfiguration = $this->getParameter('oai');
         $identify = $this->oaiService->getIdentify($url, $oaiConfiguration);
-
         $response = new Response();
         $response->setContent($this->serializer->serialize($identify, 'xml'));
         $response->headers->add(['Content-Type' => 'application/xml']);
@@ -39,9 +30,7 @@ class OaiController extends AbstractController
         return $response;
     }
 
-    /**
-     * @Route("/oai2/")
-     */
+    #[Route(path: '/oai2/')]
     public function indexAction(Request $request): Response
     {
         if ('ListSets' === $request->get('verb')) {
@@ -53,27 +42,21 @@ class OaiController extends AbstractController
         if ('Identify' === $request->get('verb')) {
             return $this->forward('Subugoe\\OaiBundle\\Controller\\OaiController::identify');
         }
-
         $response = new Response();
         $response->setContent($this->oaiService->start());
         $response->headers->add(['Content-Type' => 'application/xml']);
         $response->setStatusCode(Response::HTTP_OK);
-
         $this->oaiService->deleteExpiredResumptionTokens();
 
         return $response;
     }
 
-    /**
-     * @Route("/oai2/verb/ListMetadataFormats")
-     */
+    #[Route(path: '/oai2/verb/ListMetadataFormats')]
     public function listMetadataFormats(Request $request)
     {
         $url = $request->getSchemeAndHttpHost().$request->getPathInfo();
         $oaiConfiguration = $this->getParameter('oai');
-
         $metadataFormats = $this->oaiService->getMetadataFormats($url, $oaiConfiguration, $request->get('identifier'));
-
         $response = new Response();
         $response->setContent($this->serializer->serialize($metadataFormats, 'xml'));
         $response->headers->add(['Content-Type' => 'application/xml']);
@@ -81,14 +64,11 @@ class OaiController extends AbstractController
         return $response;
     }
 
-    /**
-     * @Route("/oai2/verb/ListSets")
-     */
+    #[Route(path: '/oai2/verb/ListSets')]
     public function listSets(Request $request)
     {
         $url = $request->getSchemeAndHttpHost().$request->getPathInfo();
         $collections = $this->getParameter('collections');
-
         $sets = $this->oaiService->getListSets($url, $collections);
         $response = new Response();
         $response->setContent($this->serializer->serialize($sets, 'xml'));
